Use SPI transactions.
Use SPI.beginTransaction() and SPI.endTransaction() to prevent other
libraries from using SPI in interrupts in the middle of our SPI
transactions (i.e. while our CS is asserted). This seems to help
prevent conflicts when using the CC3000 with this display.

Only applies to hardware SPI and only if the Arduino SPI library
supports it (SPI_HAS_TRANSACTION is defined). Applies to both AVR and
SAM (Arduino Due).

Requires an #include <SPI.h> in Adafruit_ST7735.h for SPISettings.