lcd: use writePattern for faster erases and lines.
The ESP8266 SPI implementation has a 'writePattern' function that
writes a repeating sequence (in this case a solid colour) to the SPI
port.  This fills up the FIFO and increases the bus utilisation.

Using this for clearing the screen cuts the demo run time from 4721 ms
to 3601 ms.  Also using this for the fast lines cuts the run time to
3424 ms.

Signed-off-by: Michael Hope <mlhx@google.com>